答えがマイナスの時にも対応するようにした

+++ backend/handler/CalcHandler.go
@@ -16,27 +16,27 @@ type CalcServiceServer struct {
 
 func (s *CalcServiceServer) Addition(ctx context.Context, r *pb.CalcRequest) (*pb.CalcResponse, error) {
 	log.Printf("Recieved : Add %d , %d", r.GetNum1(), r.GetNum2())
-	return &pb.CalcResponse{Result: r.GetNum1() + r.GetNum2()}, nil
+	return &pb.CalcResponse{Result: int32(r.GetNum1() + r.GetNum2())}, nil
 }
 
 func (s *CalcServiceServer) Division(ctx context.Context, r *pb.CalcRequest) (*pb.CalcResponse, error) {
 	log.Printf("Recieved : Divide %d , %d", r.GetNum1(), r.GetNum2())
-	return &pb.CalcResponse{Result: r.GetNum1() / r.GetNum2()}, nil
+	return &pb.CalcResponse{Result: int32(r.GetNum1() / r.GetNum2())}, nil
 }
 
 func (s *CalcServiceServer) Multiplication(ctx context.Context, r *pb.CalcRequest) (*pb.CalcResponse, error) {
 	log.Printf("Recieved : Multiply %d , %d", r.GetNum1(), r.GetNum2())
-	return &pb.CalcResponse{Result: r.GetNum1() * r.GetNum2()}, nil
+	return &pb.CalcResponse{Result: int32(r.GetNum1() * r.GetNum2())}, nil
 }
 
 func (s *CalcServiceServer) Subtraction(ctx context.Context, r *pb.CalcRequest) (*pb.CalcResponse, error) {
 	log.Printf("Recieved : Subtract %d , %d", r.GetNum1(), r.GetNum2())
-	return &pb.CalcResponse{Result: r.GetNum1() - r.GetNum2()}, nil
+	return &pb.CalcResponse{Result: int32(r.GetNum1() - r.GetNum2())}, nil
 }
 
 func (s *CalcServiceServer) Remain(ctx context.Context, r *pb.CalcRequest) (*pb.CalcResponse, error) {
 	log.Printf("Recieved : Remain %d, %d", r.GetNum1(), r.GetNum2())
-	return &pb.CalcResponse{Result: r.GetNum1() % r.GetNum2()}, nil
+	return &pb.CalcResponse{Result: int32(r.GetNum1() % r.GetNum2())}, nil
 }
 
 func (s *CalcServiceServer) AuthFuncOverride(ctx context.Context, fullMethodName string) (context.Context, error) {
